requirements.txt
discord.py==1.2.2
async-timeout==3.0.1
environs==4.2.0